RavenDB-22561 PR comments
